% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pse_get.R
\name{pse_get}
\alias{pse_get}
\title{Pull Historical PSE Data}
\usage{
pse_get(codes)
}
\arguments{
\item{codes}{Vector of Stock Code(s) from PSE (e.g. JFC, FMETF, BPI)}
}
\value{
A dataframe with available historical stock data
}
\description{
This function extracts historical PSE data on close, high, low, and open
values from a source. Currently only pulls from Investagrams and limited
to past one year data.
}
\examples{
pse_get("JFC")
}
