% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_pse_data.R
\name{pull_pse_data}
\alias{pull_pse_data}
\title{Pull Historical PSE Data}
\usage{
pull_pse_data(code)
}
\arguments{
\item{code}{Stock Code from PSE (e.g. JFC, FMETF, BPI)}
}
\value{
A dataframe with available historical stock data
}
\description{
This function extracts historical PSE data on close, high, low, and open
values from a source. Currently only pulls from Investagrams and limited
to past one year data.
}
\examples{
pull_pse_data("JFC")
}
