% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_investagram.R
\name{scrape_investagram}
\alias{scrape_investagram}
\title{Scrape Investagram Stock Info}
\usage{
scrape_investagram(code, details = c("hist"))
}
\arguments{
\item{code}{Vector of Stock Code(s) from PSE (e.g. JFC, FMETF, BPI)}

\item{details}{Vector containing details to scrape. `hist` - dataframe of historical data}
}
\value{
A list with investagram information
}
\description{
This function extracts raw HTML from Investagram UI and 
returns list of details
}
\examples{
scrape_investagram(c("JFC", "GLO"))
}
